refactor: util 폴더, 함수 분리 - VideoPlayer.jsx, VideoUpload.jsx를 util 폴더로 이동 - formatTime 함수 util파일 내부로 이동
